@@ -1,13 +1,12 @@
-"use client"
-import Image from "next/image";
-import { useRouter } from "next/router";
+"use client";
+import { useRouter } from "next/navigation";
 import { useEffect } from "react";
 
 export default function Home() {
   const router = useRouter();
   useEffect(() => {
     router.push("/auth/register")
-  }, [])
+  }, [router])
   
   return (
     <main>
