@@ -1,18 +1,14 @@
-import { getCourse } from '@/app/lib/getCourse';
-import { MainCourse } from '@/app/utils/interface';
-import React from 'react';
-import CheckOutPageClient from '../../../components/CheckOutPageClient';
+import CheckOutPageClient from "@/app/components/CheckOutPageClient";
+import { getCourse } from "@/app/lib/getCourse";
+import { MainCourse } from "@/app/utils/interface";
 
-type checkPageProps = {
-   params: { slug: string };
-};
+export default async function CheckOutPage({ params }: { params: Promise<{ slug: string }> }) {
+  const resolvedParams = await params; // Await the Promise if required
+  const course: MainCourse | null = await getCourse(resolvedParams.slug);
 
-export default async function CheckOutPage({ params }: checkPageProps) {
-    const course: MainCourse | null = await getCourse(params.slug);
+  if (!course) {
+    return <p>Course not found</p>;
+  }
 
-    if (!course) {
-        return <p>Course not found</p>;
-    }
-
-    return <CheckOutPageClient course={course} />;
+  return <CheckOutPageClient course={course} />;
 }
