@@ -1,8 +1,6 @@
 
 
-type Props = {}
-
-const page = (props: Props) => {
+const page = () => {
   return (
     <div>
         
