@@ -1,6 +1,7 @@
 import React from 'react'
 import { MainCourse } from '../utils/interface'
-import { handlePayment } from '../lib/paymentHandle'
+import { HandlePayment } from '../lib/paymentHandle'
+import { useSession } from 'next-auth/react'
 
 type CheckoutProps = {
   course: MainCourse | null,
@@ -9,11 +10,18 @@ type CheckoutProps = {
 }
 
 const CheckoutBtn = ({course, gateway, checkProceed}:CheckoutProps) => {
+  const { data: session } = useSession();
+
+  const handlePaymentClick = () => {
+      if (!session) return;
+      HandlePayment({ gateway, course, session });
+  };
+
   return (
     <div className="w-full mt-7 flex items-center justify-center sora ">
       <button 
       className="w-[90%] rounded-[15px] p-3 bg-[#DB0D0D] text-center text-white text-[12px]/[15.12px] font-normal disabled:bg-[#AC8D8A] disabled:text-red-500"
-      onClick={()=> handlePayment({gateway, course})}
+      onClick={handlePaymentClick}
       disabled={!checkProceed} 
       >
           Procced to pay 
