@@ -1,5 +1,4 @@
 import React from 'react';
-import { Lecture } from '../utils/interface';
 import lockIcon from '../../public/lock-circle.svg';
 import playIcon from '../../public/play-circle.svg';
 import docuIcon from '../../public/document-text.svg';
@@ -9,12 +8,12 @@ import Link from 'next/link';
 type videoPillProps = {
   btnType: "locked"  | "open" | "document",
   mainText?: string,
-  altText: any,
+  altText: string | undefined,
   proceed: boolean,
   docuLink?: string
 }
 
-export const VideoContent = ({btnType, mainText, altText, proceed,docuLink}: videoPillProps) =>{
+export const VideoContent = ({btnType, mainText, altText, proceed}: videoPillProps) =>{
   return(
     <button className='w-full rounded-xl flex gap-1 bg-[#1A1B1A] text-white items-center sora p-3 ' disabled={proceed ? true : false} >
       <Image src={btnType === "locked" ? lockIcon : btnType === "open" ? playIcon : docuIcon } 
