@@ -1,4 +1,4 @@
-import { Inter, Sora } from 'next/font/google';
+import { Sora } from 'next/font/google';
 export const sora = Sora({ 
     weight: ['400','600',],
     subsets: ['latin'] 
